* jc/combine:
  stripspace: make sure not to leave an incomplete line.
  git-commit: do not muck with commit message when no_edit is set.
  When showing a commit message, do not lose an incomplete line.
  Retire t5501-old-fetch-and-upload test.
  combine-diff: type fix.

+++ combine-diff.c
@@ -506,8 +506,8 @@ static void dump_sline(struct sline *sline, unsigned long cnt, int num_parent)
 
 	while (1) {
 		struct sline *sl = &sline[lno];
-		int hunk_end;
-		int rlines;
+		unsigned long hunk_end;
+		unsigned long rlines;
 		while (lno <= cnt && !(sline[lno].flag & mark))
 			lno++;
 		if (cnt < lno)
